requirements.txt
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
Click==7.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
fake-useragent==0.1.11
Flask==1.1.1
future==0.18.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python-dotenv==0.10.5
python-telegram-bot==12.3.0
queuelib==1.5.0
Scrapy==1.8.0
scrapy-fake-useragent==1.2.0
service-identity==18.1.0
six==1.14.0
SQLAlchemy==1.3.13
tornado==6.0.3
Twisted==19.10.0
w3lib==1.21.0
Werkzeug==0.16.0
zope.interface==4.7.1
